requirements.txt
python-dotenv
datasets
huggingface_hub
transformers
peft
gdown